+++ CHANGELOG.md
@@ -3,9 +3,30 @@ duosecurity CHANGELOG
 
 This file is used to list changes made in each version of the duosecurity cookbook.
 
-0.1.0
+1.4.1
+----
+- [mattlqx] - Remove expired Duo repo GPG key and re-add.
+
+1.4.0
+----
+- [mattlqx] - Compatibility for Ubuntu 18.04 Bionic Beaver.
+
+1.3.5
 -----
-- [your_name] - Initial release of duosecurity
+- [mattlqx] - Bump to latest duo_unix version.
+- [whiteley] - Change metadata to support Chef 13.
+
+1.3.4
+-----
+- [whiteley] - bump to latest duo_unix version
+
+1.3.3
+-----
+- [mattlqx] - Add `issues_url`, `chef_version` to metadata.rb. Taken ownership of cookbook. 🤹🏻
+
+1.3.2
+-----
+- [mattlqx] - Fix GPG keys for Duo Security package repo
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.